program: Dont hardcode the endpoint address but find the first out endpoint
Some of the older and newer tags use different USB chips und the same brand.
They differentiate in usb descriptor configuration which is causing the
programm command to fail as the descriptor is never found.

This is fixing it by using the lambda as proposed by @ngkz to find the
first valid OUT descriptor.
https://github.com/jnweiger/led-name-badge-ls32/issues/41#issuecomment-1786602203

Signed-off-by: Mimoja <git@mimoja.de>